# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Family Hub API",
    "version": "0.1.0",
    "description": "Multi-tenant family-coordination API. This spec is generated from the live Hono route table (FHS-356) — every endpoint here is a real, mounted route."
  },
  "servers": [
    {
      "url": "/",
      "description": "current host"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase user JWT (parent / adult sign-in)."
      },
      "kidAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "kid-JWT",
        "description": "Kid PIN session token from POST /api/auth/kid-pin."
      }
    }
  },
  "paths": {
...
    "/api/assignments": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "Get /api/assignments",
        "operationId": "get_api_assignments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "assignments"
        ],
        "summary": "Create /api/assignments",
        "operationId": "post_api_assignments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/auth/kid-pin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a kid PIN for a kid session token",
        "operationId": "post_api_auth_kid_pin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenantSlug": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 63,
                    "pattern": "^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
                  },
                  "memberId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "pin": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "tenantSlug",
                  "memberId",
                  "pin"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "member": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "avatarEmoji": {
                          "type": "string",
                          "nullable": true
                        },
                        "tenantId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "tenantSlug": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "displayName",
                        "avatarEmoji",
                        "tenantId",
                        "tenantSlug"
                      ],
                      "additionalProperties": false
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "token",
                    "member",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "The family's calendar activities for a week (?weekStart=YYYY-MM-DD)",
        "operationId": "get_api_events",
        "parameters": [
          {
            "name": "weekStart",
            "in": "query",
            "required": true,
            "description": "Monday of the week to fetch (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "One entry per occurrence in the 7-day window. A weekly-recurring series (recurrenceDays set) contributes one entry per matching weekday, all sharing the same series id, with isRecurring: true",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "weekStart": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "date": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "startTime": {
                            "type": "string",
                            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                            "nullable": true
                          },
                          "endTime": {
                            "type": "string",
                            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                            "nullable": true
                          },
                          "title": {
                            "type": "string"
                          },
                          "notes": {
                            "type": "string",
                            "nullable": true
                          },
                          "memberId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "school",
                              "home"
                            ]
                          },
                          "location": {
                            "type": "string",
                            "nullable": true
                          },
                          "wear": {
                            "type": "string",
                            "nullable": true
                          },
                          "recurrenceDays": {
                            "type": "array",
                            "items": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 6
                            },
                            "nullable": true
                          },
                          "recurrenceEndDate": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                            "nullable": true
                          },
                          "isRecurring": {
                            "type": "boolean"
                          },
                          "seriesStartDate": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          }
                        },
                        "required": [
                          "id",
                          "date",
                          "startTime",
                          "endTime",
                          "title",
                          "notes",
                          "memberId",
                          "type",
                          "location",
                          "wear",
                          "recurrenceDays",
                          "recurrenceEndDate",
                          "isRecurring",
                          "seriesStartDate"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "weekStart",
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create a calendar activity; admin/adult only",
        "operationId": "post_api_events",
        "description": "Optional recurrenceDays (weekdays 0=Sun..6=Sat) + recurrenceEndDate turn it into a weekly-repeating series anchored on `date`. No occurrence rows are stored — GET expands the series on read.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "startTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                    "nullable": true
                  },
                  "endTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                    "nullable": true
                  },
                  "memberId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "school",
                      "home"
                    ],
                    "default": "home"
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  },
                  "wear": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  },
                  "recurrenceDays": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 6
                    },
                    "minItems": 1,
                    "maxItems": 7,
                    "nullable": true
                  },
                  "recurrenceEndDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "nullable": true
                  }
                },
                "required": [
                  "date",
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "201 — the created event (the series anchor, if recurring)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "date": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    },
                    "startTime": {
                      "type": "string",
                      "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                      "nullable": true
                    },
                    "endTime": {
                      "type": "string",
                      "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    },
                    "memberId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "school",
                        "home"
                      ]
                    },
                    "location": {
                      "type": "string",
                      "nullable": true
                    },
                    "wear": {
                      "type": "string",
                      "nullable": true
                    },
                    "recurrenceDays": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 6
                      },
                      "nullable": true
                    },
                    "recurrenceEndDate": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                      "nullable": true
                    },
                    "isRecurring": {
                      "type": "boolean"
                    },
                    "seriesStartDate": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    }
                  },
                  "required": [
                    "id",
                    "date",
                    "startTime",
                    "endTime",
                    "title",
                    "notes",
                    "memberId",
                    "type",
                    "location",
                    "wear",
                    "recurrenceDays",
                    "recurrenceEndDate",
                    "isRecurring",
                    "seriesStartDate"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit a beta survey response",
        "operationId": "post_api_feedback",
        "description": "Authenticated users submit product feedback. All fields are optional but at least one must be present.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pmfDisappointment": {
                    "type": "string",
                    "enum": [
                      "very",
                      "somewhat",
                      "not"
                    ]
                  },
                  "recommendScore": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10
                  },
                  "solvesProblem": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "easeOfUse": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "keepUsing": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "painPoint": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "featureRequest": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "otherFeedback": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "201 — the new feedback row id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "success",
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitations": {
      "post": {
        "tags": [
          "invitations"
        ],
        "summary": "Invite someone to the family",
        "operationId": "post_api_invitations",
        "description": "role is one of admin | adult | teen | guest (default adult). Granting 'admin' requires the caller to already be an admin (403 otherwise) — FHS-486 / ADR 0019.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "adult",
                      "teen",
                      "guest"
                    ],
                    "default": "adult"
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invitation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "tenantId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "tenantId",
                        "email",
                        "role",
                        "status",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "invitation"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/{id}/resend": {
      "post": {
        "tags": [
          "invitations"
        ],
        "summary": "Create /api/invitations/{id}/resend",
        "operationId": "post_api_invitations_id_resend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/invitations/claim": {
      "post": {
        "tags": [
          "invitations"
        ],
        "summary": "Claim invites addressed to my email",
        "operationId": "post_api_invitations_claim",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}